refactor(geoApi): create common function module
